Update: Scraper code now in separate file.
